Split tensornetwork.py into network and network_components.
