+++ tensornetwork/__init__.py
@@ -1,4 +1,5 @@
 from __future__ import absolute_import
-from tensornetwork.tensornetwork import TensorNetwork, Node, Edge
+from tensornetwork.network import TensorNetwork
+from tensornetwork.network_components import Node, Edge
 from tensornetwork.ncon_interface import ncon, ncon_network
 from tensornetwork import contractors
